refactor(projective-grid): co-locate homography estimators into geometry/
Phase 1 of the projective-grid structural cleanup. The crate had two
homography code paths: geometry::estimate_projective (plain DLT ->
Projective2, used for the lattice fit + consistency check) and a separate
detect/advanced/square/homography module (Hartley-normalised DLT + 4-point
solver + HomographyQuality, used for local per-cell fits in seed, validate,
and extension, plus one cross-crate caller in calib-targets-chessboard).

Move the second module to geometry/homography.rs and re-export its public
items from geometry, so all homography estimation lives in one place. This
is a pure relocation: no numerics change, both estimators are byte-for-byte
identical to before. Callers are repointed from
projective_grid::detect::advanced::square::homography to
projective_grid::geometry.

Verified: cargo clippy --workspace --all-targets -D warnings (clean),
projective-grid + calib-targets-chessboard tests pass, cargo doc --no-deps
zero warnings.
