+++ crates/projective-grid/benches/homography.rs
@@ -4,7 +4,7 @@
 
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use nalgebra::{Matrix3, Point2};
-use projective_grid::detect::advanced::square::homography::{
+use projective_grid::geometry::{
     estimate_homography, estimate_homography_with_quality, homography_from_4pt,
     homography_from_4pt_with_quality, Homography,
 };
