+++ crates/projective-grid/src/detect/advanced/square/extension/global.rs
@@ -13,7 +13,7 @@ use nalgebra::Point2;
 use crate::detect::advanced::square::extension::common::{try_attach_at_cell, TryCellResult};
 use crate::detect::advanced::square::extension::{ExtensionParams, ExtensionStats};
 use crate::detect::advanced::square::grow::{GrowResult, SquareAttachPolicy};
-use crate::detect::advanced::square::homography::estimate_homography_with_quality;
+use crate::geometry::estimate_homography_with_quality;
 
 /// Try to extend the labelled grid outward (and into interior holes)
 /// using a globally-fit homography. Mutates `grow.labelled` and
