+++ crates/projective-grid/src/detect/advanced/square/validate/local_h.rs
@@ -4,7 +4,7 @@
 //! in `(i, j)`-space, fits a 4-point local homography and measures the
 //! reprojection residual.
 
-use crate::detect::advanced::square::homography::homography_from_4pt;
+use crate::geometry::homography_from_4pt;
 use crate::detect::advanced::square::validate::LabelledEntry;
 use nalgebra::Point2;
 use std::collections::HashMap;
